+++ dialer_other.go
@@ -1,0 +1,5 @@
+//go:build !linux
+
+package client
+
+func applySocketOptions(fd uintptr, config *Config) {}